✨完善考试题目生成、保存与查询 ✨新增用户信息查询功能 🎨优化代码鉴权，使用注解进行角色验证 🎨清理代码，去除多余import 📝更新文档说明，包括数据库建表

+++ src/main/java/com/losgai/ai/service/exam/impl/TestServiceImpl.java
@@ -62,6 +62,12 @@ public class TestServiceImpl implements TestService {
     @Override
     @Description("删除考试信息")
     public ResultCodeEnum delete(Long id) {
+        Test test = testMapper.selectByPrimaryKey(id);
+        Date now = Date.from(Instant.now());
+        // 不删除正在进行的考试
+        if (now.after(test.getStartTime()) && now.before(test.getEndTime())) {
+            return ResultCodeEnum.TIMING_ERROR;
+        }
         testMapper.deleteByPrimaryKey(id);
         return ResultCodeEnum.SUCCESS;
     }
